Support skipping tests instead of hiding them
It's easy to lose track of the tests that have been disabled due to
filtering. To help with that problem, support skipping tests instead of
hiding them completely. Expose this via an additional `--skip-filtered`
CLI option. If set, it will skip filtered tests. If unset (the default),
it will hide filtered tests like it does today.

Resolve #17

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>